added height:0 to tab container
What happened was after it replaced the images on the hidden tabs
(display: none;) it tried to set the width and height image attributes
to offsetWidth and offsetHeight of the original @1x image respectively.
Since the image was originally in a parent element with display: none;
those values were 0. By setting the height of my hidden tabs to 0 and
overflow: hidden; instead of display: none;  galengidman